Move SlfLoggingSpanCollector to com.github.kristofa.brave.ext package
